app/requirements/local.txt
asgiref==3.4.1
astroid==2.6.6
backports.entry-points-selectable==1.1.0
cfgv==3.3.0
distlib==0.3.2
Django==3.2.6
django-debug-toolbar==3.2.1
filelock==3.0.12
identify==2.2.13
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
nodeenv==1.6.0
platformdirs==2.2.0
pre-commit==2.14.0
pylint==2.9.6
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.7.2
wrapt==1.12.1
django-simple-history==3.0.0
psycopg2-binary==2.9
gunicorn==20.1.0
pyodbc==4.0.30
coverage==6.4.1
factory_boy==3.2.1
celery==5.1.2
django-celery-beat==2.2.0
django-celery-results==2.2.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
